refactor the validator into an object rather than just some class methods
this is a lot more of a ruby style interface, and the validator object behave like
the results object as well. this maintains backwards compat with the existing API
but also exposes a non-prima dona method to access the results without a rescue.

additionally this switches to a throw/catch for flow control instead of exceptions
because exceptions for flow control in ruby is typically a code smell and anti-pattern
and slower than throw/catch
see http://mensfeld.pl/2015/03/exceptions-should-not-be-expected-stop-using-them-for-control-flow-or-any-other-logic-handling-in-ruby/